+++ src/main/java/edu/example/dev_2_cc/api_controller/AdminController.java
@@ -2,8 +2,14 @@ package edu.example.dev_2_cc.api_controller;
 
 import edu.example.dev_2_cc.dto.member.MemberResponseDTO;
 import edu.example.dev_2_cc.dto.member.MemberUpdateDTO;
+import edu.example.dev_2_cc.dto.order.OrderResponseDTO;
+import edu.example.dev_2_cc.dto.order.OrderUpdateDTO;
+import edu.example.dev_2_cc.entity.Orders;
 import edu.example.dev_2_cc.service.MemberService;
+import edu.example.dev_2_cc.service.OrderService;
 import org.springframework.validation.annotation.Validated;
+
+import java.util.HashMap;
 import java.util.List;
 import edu.example.dev_2_cc.dto.product.ProductRequestDTO;
 import edu.example.dev_2_cc.dto.product.ProductResponseDTO;
@@ -24,8 +30,9 @@ import java.util.Map;
 public class AdminController {
     private final ProductService productService;
     private final MemberService memberService;
+    private final OrderService orderService;
   
-    // 회원 리관리
+    // 회원 관리
     // 관리자의 회원 전체 조회
     @GetMapping("/memberlist")
     public ResponseEntity<List<MemberResponseDTO>> getAllMembers() {
@@ -76,7 +83,54 @@ public class AdminController {
         productService.delete(productId);
         return ResponseEntity.ok(Map.of("message", "Product deleted"));
     }
-    
+
+    //주문 관리
+    //주문 전체 조회
+    @GetMapping("/order")
+    public ResponseEntity<List<OrderResponseDTO>> getAllOrders() {
+        List<OrderResponseDTO> orderList = orderService.list();
+        return ResponseEntity.ok(orderList);
+
+    }
+
+    //단일 주문 조회
+    @GetMapping("/order/{orderId}")
+    public ResponseEntity<OrderResponseDTO> getOrder(
+            @PathVariable Long orderId) {
+        Orders order = orderService.findOrderById(orderId);
+        OrderResponseDTO orderResponseDTO = new OrderResponseDTO(order);
+
+        return ResponseEntity.ok(orderResponseDTO);
+    }
+
+    //배송 상태 수정
+    @PutMapping("/order/{orderId}")
+    public ResponseEntity<OrderResponseDTO> updateOrder(
+            @PathVariable Long orderId,
+            @RequestBody OrderUpdateDTO orderUpdateDTO) {
+
+        //updateDTO에 orderId 설정
+        orderUpdateDTO.setOrderId(orderId);
+        //배송 상태 수정
+        Orders order = orderService.modifyStatus(orderUpdateDTO);
+        OrderResponseDTO orderResponseDTO = new OrderResponseDTO(order);
+
+        return ResponseEntity.ok(orderResponseDTO);
+
+    }
+
+    //주문 삭제
+    @DeleteMapping("/order/{orderId}")
+    public ResponseEntity<Map<String, String>> deleteOrder(
+            @PathVariable Long orderId) {
+        orderService.delete(orderId); // 서비스에서 주문 삭제 로직 호출
+        // 메시지를 담은 응답
+
+        Map<String, String> response = new HashMap<>();
+        response.put("message", "주문이 성공적으로 삭제되었습니다.");
+
+        return ResponseEntity.ok(response); // 200 OK와 함께 응답 본문 반환
+    }
 
     
 }
