+++ src/main/java/edu/example/dev_2_cc/api_controller/MypageController.java
@@ -3,9 +3,13 @@ package edu.example.dev_2_cc.api_controller;
 
 import edu.example.dev_2_cc.dto.member.MemberResponseDTO;
 import edu.example.dev_2_cc.dto.member.MemberUpdateDTO;
+import edu.example.dev_2_cc.dto.order.OrderRequestDTO;
+import edu.example.dev_2_cc.dto.order.OrderResponseDTO;
+import edu.example.dev_2_cc.entity.Orders;
 import edu.example.dev_2_cc.exception.MemberException;
 import edu.example.dev_2_cc.exception.MemberTaskException;
 import edu.example.dev_2_cc.service.MemberService;
+import edu.example.dev_2_cc.service.OrderService;
 import org.springframework.http.HttpStatus;
 import edu.example.dev_2_cc.dto.review.ReviewRequestDTO;
 import edu.example.dev_2_cc.dto.review.ReviewResponseDTO;
@@ -16,6 +20,8 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.HashMap;
 import java.util.Map;
 
 @RestController
@@ -25,6 +31,7 @@ import java.util.Map;
 public class MypageController {
     private final MemberService memberService;
     private final ReviewService reviewService;
+    private final OrderService orderService;
 
     // 회원 정보 수정
     // 마이페이지 내에서 회원의 직접 정보 수정 (권한 수정 미포함)
@@ -56,6 +63,40 @@ public class MypageController {
         }
     }
 
+    //주문
+    //주문 생성
+    @PostMapping("/order")
+    public ResponseEntity<OrderResponseDTO> createOrder(
+            @RequestBody OrderRequestDTO orderRequestDTO) {
+        Orders order = orderService.createOrder(orderRequestDTO);
+        OrderResponseDTO orderResponseDTO = new OrderResponseDTO(order);
+
+        return ResponseEntity.ok(orderResponseDTO);
+    }
+
+    //단일 주문 조회
+    @GetMapping("/order/{orderId}")
+    public ResponseEntity<OrderResponseDTO> getOrder(
+            @PathVariable Long orderId) {
+        Orders order = orderService.findOrderById(orderId);
+        OrderResponseDTO orderResponseDTO = new OrderResponseDTO(order);
+
+        return ResponseEntity.ok(orderResponseDTO);
+    }
+
+    //주문 삭제
+    @DeleteMapping("/order/{orderId}")
+    public ResponseEntity<Map<String, String>> deleteOrder(
+            @PathVariable Long orderId) {
+        orderService.delete(orderId); // 서비스에서 주문 삭제 로직 호출
+        // 메시지를 담은 응답
+
+        Map<String, String> response = new HashMap<>();
+        response.put("message", "주문이 성공적으로 삭제되었습니다.");
+
+        return ResponseEntity.ok(response); // 200 OK와 함께 응답 본문 반환
+    }
+
     //리뷰
     @PostMapping("/review")
     public ResponseEntity<ReviewResponseDTO> createReview(@RequestBody ReviewRequestDTO reviewRequestDTO) {
@@ -72,5 +113,7 @@ public class MypageController {
         reviewService.delete(reviewId);
         return ResponseEntity.ok(Map.of("message", "Review deleted"));
     }
+
+
 }
